fix(ui): set custom theme in twmerge config
