+++ packages/tsconfig/ui.json
@@ -1,7 +1,6 @@
 {
     "extends": "./base.json",
     "compilerOptions": {
-        "jsx": "react",
-        "allowJs": true
+        "jsx": "react"
     }
 }
