+++ packages/ui/src/tailwind.preset.theme.d.ts
@@ -1,0 +1,3 @@
+import { Config } from "tailwindcss";
+
+declare const theme: Config["theme"];
